+++ src/main/java/com/game/monopoly/controller/ChatController.java
@@ -3,7 +3,7 @@ package com.game.monopoly.controller;
 import com.game.monopoly.dto.request.CommonMessageDTO;
 import com.game.monopoly.dto.response.ResultMessageDTO;
 import com.game.monopoly.dto.response.SuccessMessageDTO;
-import com.game.monopoly.service.SessionService;
+import com.game.monopoly.service.ChatService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.handler.annotation.MessageMapping;
@@ -15,12 +15,12 @@ import static com.game.monopoly.constants.ResultMessage.COMMON_MESSAGE_WAS_SAVED
 @Controller
 @RequiredArgsConstructor
 public class ChatController {
-    private final SessionService sessionService;
+    private final ChatService chatService;
     private final SimpMessagingTemplate simpMessagingTemplate;
 
     @MessageMapping(value = "/chat/common")
     public ResponseEntity<SuccessMessageDTO> saveCommonMessage(CommonMessageDTO dto) {
-        sessionService.addCommonMessageToChatHistory(dto.getSessionId(), dto.getSender(), dto.getMessage());
+        chatService.addCommonMessageToChatHistory(dto.getSessionId(), dto.getSender(), dto.getMessage());
         simpMessagingTemplate.convertAndSend(
                 "/topic/chat/" + dto.getSessionId(),
                 new ResultMessageDTO(dto.getSender(), dto.getMessage())
