+++ src/main/java/com/game/monopoly/controller/CompanyCardController.java
@@ -1,7 +1,7 @@
 package com.game.monopoly.controller;
 
 import com.game.monopoly.dto.response.CardDetailDTO;
-import com.game.monopoly.service.SessionService;
+import com.game.monopoly.service.CompanyCardService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,13 +13,13 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1")
 @RequiredArgsConstructor
 public class CompanyCardController {
-    private final SessionService sessionService;
+    private final CompanyCardService companyCardService;
 
     @GetMapping(value = "/cards", params = {"sessionId", "cardId"})
     public ResponseEntity<CardDetailDTO> getDetailedCard(
             @RequestParam("sessionId") String sessionId,
             @RequestParam("cardId") Long cardId) {
-        CardDetailDTO result = sessionService.getDetailedCardInfo(sessionId, cardId);
+        CardDetailDTO result = companyCardService.getDetailedCardInfo(sessionId, cardId);
 
         return ResponseEntity.ok().body(result);
     }
