+++ src/main/java/com/game/monopoly/controller/SessionProcessController.java
@@ -1,32 +1,34 @@
 package com.game.monopoly.controller;
 
-import com.game.monopoly.dto.request.*;
+import com.game.monopoly.dto.request.ChangeCurrentPlayerDTO;
+import com.game.monopoly.dto.request.InitializeSessionDTO;
+import com.game.monopoly.dto.request.SessionIdDTO;
+import com.game.monopoly.dto.request.SessionPlayerNameDTO;
 import com.game.monopoly.dto.response.*;
 import com.game.monopoly.entity.Player;
 import com.game.monopoly.enums.MoveStatus;
 import com.game.monopoly.mapper.MessageContentMapper;
 import com.game.monopoly.mapper.PlayerMapper;
-import com.game.monopoly.service.SessionService;
+import com.game.monopoly.service.SessionProcessService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 
-import static com.game.monopoly.constants.ResultMessage.*;
+import static com.game.monopoly.constants.ResultMessage.GAME_WAS_STARTED;
+import static com.game.monopoly.constants.ResultMessage.NEW_PLAYER;
 import static com.game.monopoly.enums.SessionState.IN_PROGRESS;
 
 @Controller
 @RequiredArgsConstructor
-public class SessionWebSocketController {
-    private final SessionService sessionService;
+public class SessionProcessController {
+    private final SessionProcessService sessionProcessService;
     private final SimpMessagingTemplate simpMessagingTemplate;
 
     @MessageMapping(value = "/sessions/add-player")
     public ResponseEntity<PlayerDTO> addPlayer(InitializeSessionDTO dto) {
-        Player player = sessionService.addPlayerToSession(dto.getSessionId(), dto.getPlayerName(), dto.getColour());
+        Player player = sessionProcessService.addPlayerToSession(dto.getSessionId(), dto.getPlayerName(), dto.getColour());
         PlayerDTO playerDTO = PlayerMapper.entityToPLayerDTO(player);
         ResultMessageDTO resultMessage = new ResultMessageDTO(dto.getPlayerName(), NEW_PLAYER);
 
@@ -38,7 +40,7 @@ public class SessionWebSocketController {
 
     @MessageMapping(value = "/sessions/roll-dice")
     public ResponseEntity<RollDiceResultDTO> getNewPlayerPosition(SessionPlayerNameDTO dto) {
-        RollDiceResultDTO rollDiceResult = sessionService.rollDices(dto.getSessionId(), dto.getPlayerName());
+        RollDiceResultDTO rollDiceResult = sessionProcessService.rollDices(dto.getSessionId(), dto.getPlayerName());
         ResultMessageDTO resultMessage = new ResultMessageDTO(
                 dto.getPlayerName(),
                 MessageContentMapper.rollDiceToMessageContent(rollDiceResult.getDigits())
@@ -52,7 +54,7 @@ public class SessionWebSocketController {
 
     @MessageMapping(value = "/sessions/start-game")
     public ResponseEntity<StartGameResultDTO> startGame(ChangeCurrentPlayerDTO dto) {
-        sessionService.startGame(dto.getSessionId(), dto.getPlayerName());
+        sessionProcessService.startGame(dto.getSessionId(), dto.getPlayerName());
 
         StartGameResultDTO result = new StartGameResultDTO(IN_PROGRESS.toString(), dto.getPlayerName());
         ResultMessageDTO resultMessage = new ResultMessageDTO(null, GAME_WAS_STARTED);
@@ -65,7 +67,7 @@ public class SessionWebSocketController {
 
     @MessageMapping(value = "/sessions/move-transition")
     public ResponseEntity<CurrentPlayerDTO> moveTransition(ChangeCurrentPlayerDTO dto) {
-        String currentPlayer = sessionService.getNextPlayer(dto.getSessionId(), dto.getPlayerName());
+        String currentPlayer = sessionProcessService.getNextPlayer(dto.getSessionId(), dto.getPlayerName());
         CurrentPlayerDTO result = new CurrentPlayerDTO(currentPlayer);
 
         simpMessagingTemplate.convertAndSend("/topic/move-transition/" + dto.getSessionId(), result);
@@ -73,31 +75,9 @@ public class SessionWebSocketController {
         return ResponseEntity.ok().body(result);
     }
 
-    @PostMapping(value = "/sessions/buy-card")
-    public ResponseEntity<BuyCardDTO> buyCard(@RequestBody PerformActionWithCardDTO dto) {
-        BuyCardDTO result = sessionService.buyCard(dto.getSessionId(), dto.getPlayerName(), dto.getCardId());
-        ResultMessageDTO resultMessage = new ResultMessageDTO(dto.getPlayerName(), BUY_CARD);
-
-        simpMessagingTemplate.convertAndSend("/topic/buy-card/" + dto.getSessionId(), result);
-        simpMessagingTemplate.convertAndSend("/topic/chat/" + dto.getSessionId(), resultMessage);
-
-        return ResponseEntity.ok().body(result);
-    }
-
-    @MessageMapping(value = "/sessions/pay-for-card")
-    public ResponseEntity<PayForCardDTO> payForCard(PerformActionWithCardDTO dto) {
-        PayForCardDTO result = sessionService.payForCard(dto.getSessionId(), dto.getPlayerName(), dto.getCardId());
-        ResultMessageDTO resultMessage = new ResultMessageDTO(dto.getPlayerName(), PAY_FOR_CARD);
-
-        simpMessagingTemplate.convertAndSend("/topic/pay-for-card/" + dto.getSessionId(), result);
-        simpMessagingTemplate.convertAndSend("/topic/chat/" + dto.getSessionId(), resultMessage);
-
-        return ResponseEntity.ok().body(result);
-    }
-
     @MessageMapping(value = "/sessions/move-status")
     public ResponseEntity<MoveStatusDTO> getCurrentMoveStatus(SessionIdDTO dto) {
-        MoveStatus status = sessionService.getCurrentMoveStatus(dto.getSessionId());
+        MoveStatus status = sessionProcessService.getCurrentMoveStatus(dto.getSessionId());
         MoveStatusDTO result = new MoveStatusDTO(String.valueOf(status));
 
         simpMessagingTemplate.convertAndSend("/topic/move-status/" + dto.getSessionId(), result);
