+++ src/main/java/com/game/monopoly/service/ChatService.java
@@ -1,0 +1,5 @@
+package com.game.monopoly.service;
+
+public interface ChatService {
+    void addCommonMessageToChatHistory(String sessionId, String sender, String message);
+}
