+++ src/main/java/com/game/monopoly/service/SessionCommonService.java
@@ -1,0 +1,9 @@
+package com.game.monopoly.service;
+
+import com.game.monopoly.entity.Session;
+
+public interface SessionCommonService {
+    Session getSession(String sessionId);
+
+    void checkSessionExists(String sessionId);
+}
