+++ src/main/java/com/game/monopoly/service/SessionProcessService.java
@@ -1,0 +1,17 @@
+package com.game.monopoly.service;
+
+import com.game.monopoly.dto.response.RollDiceResultDTO;
+import com.game.monopoly.entity.Player;
+import com.game.monopoly.enums.MoveStatus;
+
+public interface SessionProcessService {
+    Player addPlayerToSession(String sessionId, String playerName, String colour);
+
+    RollDiceResultDTO rollDices(String sessionId, String playerName);
+
+    void startGame(String sessionId, String nextPlayer);
+
+    String getNextPlayer(String sessionId, String nextPlayer);
+
+    MoveStatus getCurrentMoveStatus(String sessionId);
+}
